Remove circular dependency in manual link
